refactor: extract SpendingLimitRow to its own component
