+++ src/routes/safe/components/Balances/SendModal/screens/SendFunds/SpendingLimitRow.tsx
@@ -1,0 +1,61 @@
+import { RadioButtons, Text } from '@gnosis.pm/safe-react-components'
+import { BigNumber } from 'bignumber.js'
+import React from 'react'
+import styled from 'styled-components'
+
+import Field from 'src/components/forms/Field'
+import Col from 'src/components/layout/Col'
+import Row from 'src/components/layout/Row'
+import { SpendingLimit } from 'src/logic/safe/store/models/safe'
+import { Token } from 'src/logic/tokens/store/model/token'
+import { fromTokenUnit } from 'src/routes/safe/components/Settings/SpendingLimit/utils'
+
+// TODO: propose refactor in safe-react-components based on this requirements
+const SpendingLimitRadioButtons = styled(RadioButtons)`
+  & .MuiRadio-colorPrimary.Mui-checked {
+    color: ${({ theme }) => theme.colors.primary};
+  }
+`
+
+interface SpendingLimitRowProps {
+  onOptionSelect: () => void
+  tokenSpendingLimit: SpendingLimit
+  selectedToken: Token
+}
+
+export const SpendingLimitRow = ({
+  onOptionSelect,
+  tokenSpendingLimit,
+  selectedToken,
+}: SpendingLimitRowProps): React.ReactElement => {
+  const availableAmount = React.useMemo(() => {
+    return fromTokenUnit(
+      new BigNumber(tokenSpendingLimit.amount).minus(tokenSpendingLimit.spent).toString(),
+      selectedToken.decimals,
+    )
+  }, [selectedToken.decimals, tokenSpendingLimit.amount, tokenSpendingLimit.spent])
+
+  return (
+    <Row margin="sm">
+      <Col between="lg" style={{ flexDirection: 'column' }}>
+        <Text size="lg">Send as</Text>
+        <Field name="txType" initialValue="multiSig">
+          {({ input: { name, value } }) => (
+            <SpendingLimitRadioButtons
+              name={name}
+              value={value || 'multiSig'}
+              onRadioChange={onOptionSelect}
+              options={[
+                { label: 'Multisig Transaction', value: 'multiSig' },
+                {
+                  label: `Spending Limit Transaction (${availableAmount} ${selectedToken.symbol})`,
+                  value: 'spendingLimit',
+                },
+              ]}
+            />
+          )}
+        </Field>
+      </Col>
+    </Row>
+  )
+}
